# DOCUMENTACION Y SCRPT/logs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmarWatt Microservices API",
    "description": "API completa para análisis energético y chatbot inteligente",
    "version": "2.0.0",
    "contact": {
      "name": "SmarWatt Team",
      "email": "dev@smarwatt.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Energy IA API Development"
    },
    {
      "url": "http://localhost:5001",
      "description": "Expert Bot API Development"
    }
  ],
  "paths": {
...
    "/api/v1/analysis/sentiment": {
      "post": {
        "summary": "POST /api/v1/analysis/sentiment",
        "description": "Endpoint empresarial para análisis de sentiment avanzado.\n    Expone el método analyze_sentiment_enterprise vía HTTP para microservicios independientes.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "OPTIONS /api/v1/analysis/sentiment",
        "description": "Endpoint empresarial para análisis de sentiment avanzado.\n    Expone el método analyze_sentiment_enterprise vía HTTP para microservicios independientes.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticación de Firebase"
      }
    }
  }
}